Loads of changes to improve behaviour with python 2.7 and in 64-bit mode.
A rather significant change is that Foundation.NSRect and Quartz.CGRect
are now aliases (because of the structure of PyObjC both are Foundation.NSRect).
Likewise for NSSize and NSPoint.

I've also added objc.array_property, objc.set_property and objc.dict_property,
but haven't fleshed out unittests for these yet.

+++ pyobjc-framework-Quartz/Modules/_coregraphics.m
@@ -111,7 +111,7 @@ m_CGWindowListCreate(PyObject* self __attribute__((__unused__)),
 	}
 
 	for (i = 0; i < len; i++) {
-		CGWindowID cur = (CGWindowID)CFArrayGetValueAtIndex(windowList, i);
+		CGWindowID cur = (CGWindowID)(NSInteger)CFArrayGetValueAtIndex(windowList, i);
 		PyObject* v = PyObjC_ObjCToPython(@encode(CGWindowID), &cur);
 		if (v == NULL) {
 			CFRelease(windowList);
@@ -150,7 +150,7 @@ createWindowList(PyObject* items)
 			CFRelease(array);
 			return NULL;
 		}
-		CFArrayAppendValue(array, (const void*)windowID);
+		CFArrayAppendValue(array, (const void*)(NSInteger)windowID);
 	}
 	Py_DECREF(seq);
 	return (CFArrayRef)array;
